Added FluentValidation

+++ BG.NET.Library.Client/src/app/components/book-list-v1/book-list-v1.component.ts
@@ -4,6 +4,7 @@ import {BooksService} from "../../services/books/books.service";
 import {GenericPaginationModel} from "../../special/genericPagination.model";
 import {ActivatedRoute, Router} from "@angular/router";
 import {PageEvent} from "@angular/material/paginator";
+import {MatDialog, MatDialogRef, MatDialogModule} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-book-list-v1',
@@ -12,7 +13,10 @@ import {PageEvent} from "@angular/material/paginator";
 })
 export class BookListV1Component implements OnInit{
   public books?: GenericPaginationModel<BookFullDto>;
-  constructor(private booksService: BooksService, private router: Router, private route: ActivatedRoute) {
+  constructor(
+    private booksService: BooksService,
+    private router: Router,
+    private route: ActivatedRoute) {
 
   }
   ngOnInit(): void {
@@ -36,14 +40,22 @@ export class BookListV1Component implements OnInit{
       })
   }
   deleteBook(id: number){
-    this.booksService.deleteBook(id).subscribe({
-      next: ()=>{
-        this.handlePaginationEvent(undefined)
-      },
-      error:(response)=>{
-        console.log("Deletion error")
-        console.log(response)
-      }
-    })
+    /*
+    let dialogRef = this.dialog.open(BookListV1Component, {
+      height: '200px',
+      width: '200px',
+    });*/
+
+    if (confirm("Do you want to delete book with id "+id+"?")){
+      this.booksService.deleteBook(id).subscribe({
+        next: ()=>{
+          this.handlePaginationEvent(undefined)
+        },
+        error:(response)=>{
+          console.log("Deletion error")
+          console.log(response)
+        }
+      })
+    }
   }
 }
